Add "undo move" functionality
This is accomplished by unlinking the boards that the move created and
reverting the pieces on the old boards to their state. Only the last
move played can be undone. In the case of the low-level functions (see
moverules.undoMove), special care must be taken to not call them with
any move other than the last played move.